requirements.txt
black==23.12.1
debugpy==1.8.0
fastapi==0.108.0
folium==0.15.1
httpx==0.26.0
isort==5.13.2
pandas==2.1.4
# psycopg2==2.9.9
pytest==7.4.4
python-dotenv==1.0.0
uvicorn==0.25.0
geopandas==0.14.2
